#309  Improve context struct
