--- catalog/context/__init__.py
@@ -1,2 +1,0 @@
-from . import category, pages, products, tags
-
